TripleO and backwards compatibility
TripleO has run with good but not perfect backwards compatibility
since creation. It's time to formalise this in a documentable and
testable fashion.

We will stop deliberately breaking backwards compatibility that is
needed by any supported OpenStack release, unless there is a PTL
signoff (which will be reserved for exceptional circumstances).

Semantic versioning will make it clear to consumers of TripleO -
whether distributors or end users - when this has happened.

Change-Id: I7b63e7860d0ea58faf46c5f5483ae9fc475dbbe9