Added structure for kilo specs
This commit adds a new template for kilo to sit separate from the juno
template. This will allow us to make changes to the template for the
kilo cycle without breaking the unit tests or having to edit all of the
juno specs (a future commit has a proposal for changes to the template).

At the time of posting this, there are no kilo specs in review, so now
would be the time to make changes to the kilo template without causing
headaches.

Change-Id: Id6fffeabcc1eb003adb479c234d3d7ee06dc3a76