Add expedited approvals policy
In general, TripleO follows the standard “2 +2” review standard,
but there are situations where we want to make an exception. This
policy is intended to document those exceptions.

It has recently come to my attention that not all reviewers are
on the same page regarding some of our review policies, so this
is my attempt to document it in a reviewable location.  Some of
this information is already on the wiki, but that is not easy for
the team to review before changes are made and is probably going
away in the future.

Since this is not a release-specific document, I've added a new
section to our specs index for cross-release policies like this.
It is based on the oslo-specs policy section.  I would anticipate
this being useful for other policy type information in the future,
especially as the wiki is being phased out.

Change-Id: Ie2d357813ebc194a94ee26464d6882a6d8abe13a