Spec: undercloud-ntp-server - Add NTP services to Undercloud
This spec describes adding an NTP server to the Undercloud in
order to provide time syncing for overcloud nodes when an external
NTP server is not available.

By adding an NTP server to the Undercloud, and making its usage
optional, we allow for proof-of-concept deployments without Internet
access. This also removes a dependency on an external resource for
CI jobs.

The NTP server will be configured by the Undercloud installation
scripts. We may want to add validation to ensure that this local
NTP server can reach a stratum-1 server, or warn that the time
may be inaccurate if no server can be reached from the Undercloud.

Change-Id: Ia6ed8e8d6ebf40892cb232e39de91630c12bba48