blueprint tripleo-routed-networks-deployment
Spec: tripleo-routed-networks-deployment - Routed Spine and Leaf

This spec covers the scope of work required to enable deployment of TripleO
in a routed network environment. A routed network design that uses a
"Spine and Leaf", or "Clos" architecture, allows for horizontal scalability
where each group of servers is inside its own set of layer 2 subnets, and
subnets are connected via routers. This allows for a larger number of total
nodes, since the broadcast, multicast, and unknown unicast (BUM traffic) is
isolated to the local layer 2 domain. Furthermore, equal-cost multipath
(ECMP) can provide better distribution of traffic between multiple uplinks
in large deployments.

There are multiple projects which will need to be enhanced to add features
for routed newtorks. The DHCP services provided by both Neutron and Ironic
Inspector need to be enhanced to provide routers for each subnet. The use
of Neutron as an IP address management (IPAM) backend will need to be
evaluated, to determine if we need features that were added to support
routed networks in Neutron. The TripleO Heat templates will need to be
refactored in order to support multiple subnets. We may need some
enhancements to os-net-config (to be determined).

This spec attempts to break down the overall goals by the problem set as
it applies to each project. Separate specs will be used
for each separate project or dependency, all depenent on this one.

This spec is related to the following:

blueprint tripleo-predictable-ctlplane-ips
blueprint tripleo-routed-networks-ironic-inspector
blueprint tripleo-routed-networks-templates

Change-Id: I72861ce8adfeff67b894cf389f7096cde8be9258