blueprint tripleo-routed-networks-ironic-inspector
Spec: TripleO Ironic Inspector changes to support routed networks

This spec implements the blueprint to implement Ironic Inspector
for routed subnets. A DHCP relay is configured on the L3 switches,
and dnsmasq replies with an IP from the correct subnet. In order
for this to work, changes will have to be made to the Puppet
scripts that configure Ironic Inspector.

Note that this spec does not cover general use of Ironic
Inspector over routed networks for tenant bare metal use. That
will be covered by another blueprint which uses Neutron
integration, which is not required here.

This is a dependency of blueprint tripleo-routed-networks-deployment

Change-Id: I3c983d66fffbdbaa03ed1692fd77c95cd6c15abf