blueprint for TripleO management of AIDE
Introduce a composable service to manage AIDE and ensure
the AIDE package is installed, create rule entries and populate a cron
job to allow a periodic check of the AIDE database.

AIDE (Advanced Intrusion Detection Environment) is a file and directory
integrity verfication system. Security Frameworks such as DISA STIG /
CSI require that AIDE be installed and configured on all Linux systems.

Change-Id: I7b84df3a51c4309b2d4856072205e6eb64402da8