Improve upgrades_tasks CI coverage with standalone for Stein
Utilizing the standalone installer to better test the upgrade
tasks within existing ci wall-time and nodes.

As part of Stein PTG discussion [1] it was decided that the
approach outlined here will be one of two main streams for
upgrades CI in S. This for testing service upgrade_tasks
and another stream for testing the workflow. That latter
workflow stream is not considered here.

[1] https://etherpad.openstack.org/p/ptg_denver_2018_tripleo_ci
Co-Authored-By: Jiri Stransky <jistr@redhat.com>, Athlan-Guyot sofer <sathlang@redhat.com>
Change-Id: Ic8a8867018c6fb866856a45a2bf472a0ed65d99b