Create ansible roles for cloud operator actions
The cli has traditionally been the only supported interface for
deploying tripleo. This however has lead to multiple implementations of
roles to perform the various tripleo actions. We should establish a
small supported set of ansible roles, modules and plugins that can be
used when writing automation around TripleO deployments.

Change-Id: I8ac853b78f57661f53f77cb185806461d7f029f2